Remove the public LabelledTextTagTextBox as base class.
